+++ CHANGES.md
@@ -24,6 +24,7 @@ Changes in CUPS v2.5b1 (TBA)
 - Added Docker support (Issue #929)
 - Added a systemd slice to the systemd services included with the scheduler
 - Added localizations for deprecated IPP attributes/options (Issue #1020)
+- Added support for specifying permissions with the `cupsFileOpen` API.
 - Updated CUPS to require TLS support - OpenSSL, GNUTLS and LibreSSL are
   supported.
 - Updated CUPS to require ZLIB.